\name{Banza}
\docType{data}
\alias{Banza}
\title{Banza MCMC output}
\description{
  This data set was generated by the function PLD_interface.Rd using a phylogeny constructed from the data of Shapiro et al. (2006).
  The command was: names_locations = c("Maui_Nui","Maui_Nui","Maui_Nui","Maui_Nui","Kauai","Kauai","Maui_Nui","Maui_Nui","Maui_Nui","Maui_Nui","Nihoa","Nihoa","Hawaii","Hawaii","Hawaii","Oahu","Oahu","Maui_Nui","Maui_Nui","Oahu","Oahu");
  Banza = PLD_interface(fileTREES=system.file("extdata","tree_Banza.nex.txt",package="phyloland"),fileDATA=system.file("extdata","locations_Banza.txt",package="phyloland"), num_step=10000,freq=10,ess_lim=500,id_filena="Banza",names_locations=names_locations) 
}
\usage{banza}
\format{An object of class ``phyloland" comprising 7 attributes:
\enumerate{
    \item  trees       1000   multiPhylo list     
    \item  locations     41   data.frame list     
    \item  tips          21   -none-     character
    \item  space         10   -none-     numeric  
    \item  index          1   -none-     list     
    \item  mcmc           8   -none-     list     
    \item  sigma_limit    2   -none-     numeric
    }}
\source{Output generated by phyloland:PLD_interface.Rd}
\references{
  Shapiro LH, Strazanac JS, Roderick GK. Molecular phylogeny of Banza (Orthoptera: Tettigoniidae), the endemic katydids of the Hawaiian Archipelago.
Mol. Phylogenet. Evol. 41(1):53-63, 2006.
}
\keyword{datasets}
