\name{phyloland-internal}
\alias{ancestor}
\alias{ancestors}
\alias{check_treel}
\alias{children}
\alias{childrentip}
\alias{converttree}
\alias{effSaSize}
\alias{ftips}
\alias{height}
\alias{internal_loc}
\alias{internal_node_height}
\alias{lconverttree}
\alias{loc_ancestor}
\alias{mcmc_phyloland}
\alias{new_migration}
\alias{plot_trees_tips}
\alias{read_loc}
\alias{read_space}
\alias{replace_tips}
\alias{sigma_upperlim}
\alias{sim_history}
\alias{sim_mig}
\alias{space_dist}
\alias{stat_phyloland}
\alias{test_tree}
\alias{treeLikeli}
\alias{treel2mig}
\alias{treel2par}
\alias{verif_arg}
\alias{reorder_treel}
\alias{reorder_treep}
\alias{break_node_height_tie}
\title{Internal phyloland Functions}
\description{
 Internal phyloland functions
}
\details{
Internal functions that are not to be called by the user.
}
\keyword{ internal }

