\name{tree_Banza}
\docType{data}
\alias{tree_Banza}
\title{Phylogenetic tree of Banza}
\description{
  This data contains the tree used to generate the file banza.RData.
}
\usage{banza}
\format{A phylogenetic tree in nexus format. Taxa names match the sample names in locations_Banza.txt.}
\source{Generated in Beast using data from Shapiro et al. (2006)}
\references{
  Shapiro LH, Strazanac JS, Roderick GK. Molecular phylogeny of Banza
(Orthoptera: Tettigoniidae), the endemic katydids of the Hawaiian Archipelago.
Mol Phylogenet Evol. 2006 Oct;41(1):53-63.
}
\keyword{datasets}
